Fully implement spawn player packet and destroy entities packet

+++ Amethyst.Api/Components/VectorF.cs
@@ -1,3 +1,3 @@
 ﻿namespace Amethyst.Api.Components;
 
-public record struct VectorF(float X, float Y, float Z);
+public record struct VectorF(double X, double Y, double Z);
